refactor: (리뷰반영) repository 저장소 인터페이스 삭제
